// front/src/lib/api/types.ts
export type StatusFinanceiro = 'Não Paga' | 'Paga';